refactor(layout): replace Sidenav components with specific SidenavProfile
