+++ src/app/(modules)/(user)/profile/layout.tsx
@@ -1,21 +1,6 @@
 import { Metadata } from 'next';
 import React from 'react';
-import Sidenav from '~/components/layout/sidenav';
-import { SettingIcon, UserIcon } from '~/components/shared/icons';
-import { PATH } from '~/constants/routes';
-
-const listNav = [
-  {
-    name: '個人情報',
-    href: PATH.PROFILE.INFO,
-    icon: UserIcon,
-  },
-  {
-    name: '設定',
-    href: PATH.PROFILE.SETTING,
-    icon: SettingIcon,
-  },
-];
+import SidenavProfile from '~/components/layout/sidenav-profile';
 
 export const metadata: Metadata = {
   title: 'プロフィール',
@@ -25,9 +10,5 @@ export const metadata: Metadata = {
 };
 
 export default function layout({ children }: { children: React.ReactNode }) {
-  return (
-    <Sidenav title="プロフィール" listNav={listNav}>
-      {children}
-    </Sidenav>
-  );
+  return <SidenavProfile>{children}</SidenavProfile>;
 }
